Cleanup, download DAODAO deps
