+++ scripts/util_base.py
@@ -6,7 +6,17 @@ parent_dir = os.path.dirname(current_dir)
 
 contracts_path = os.path.join(parent_dir, "contracts")
 
-contracts_json_path = os.path.join(parent_dir, "configs", ".contracts.json")
+contracts_json_path = os.path.join(parent_dir, "configs", "contracts.json")
+
+
+def default_contracts_json():
+    if not os.path.exists(contracts_json_path):
+        with open(contracts_json_path, "w") as f:
+            f.write(json.dumps({"start_time": 0, "file_cache": {}}))
+
+
+default_contracts_json()
+
 
 # create contracts folder if not already
 if not os.path.exists(contracts_path):
